Character creation fix and concurrency fix
For real this time... Accidentally merged somehow.

CREATE packet had incorrect datatypes, and added a lock to SendToServer
and SendToClient functions.

+++ Lib K Relay/Networking/Packets/Client/CreatePacket.cs
@@ -8,22 +8,22 @@ namespace Lib_K_Relay.Networking.Packets.Client
 {
     public class CreatePacket : Packet
     {
-        public byte ClassType;
-        public byte SkinType;
+        public ushort ClassType;
+        public ushort SkinType;
 
         public override PacketType Type
         { get { return PacketType.CREATE; } }
 
         public override void Read(PacketReader r)
         {
-            ClassType = r.ReadByte();
-            SkinType = r.ReadByte();
+            ClassType = r.ReadUInt16();
+            SkinType = r.ReadUInt16();
         }
 
         public override void Write(PacketWriter w)
         {
-            w.Write(ClassType);
-            w.Write(SkinType);
+            w.Write((ushort)ClassType);
+            w.Write((ushort)SkinType);
         }
     }
 }
